// packages/components/layer-vue/enums.ts
enum Directions {
  t = 'layer-vue-resize-t',
  rt = 'layer-vue-resize-rt',
  r = 'layer-vue-resize-r',
  rb = 'layer-vue-resize-rb',
  b = 'layer-vue-resize-b',
  lb = 'layer-vue-resize-lb',
  l = 'layer-vue-resize-l',
  lt = 'layer-vue-resize-lt',
}

enum Offsets {
  t = 't',
  r = 'r',
  b = 'b',
  l = 'l',
  auto = 'auto',
  lt = 'lt',
  lb = 'lb',
  rt = 'rt',
  rb = 'rb',
}

enum CssLength {
  px = 'px',
  vw = 'vw',
  vh = 'vh',
  '%' = '%',
}
export { Directions, Offsets, CssLength };